% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findmarkers_gene2cell_mapping}
\alias{findmarkers_gene2cell_mapping}
\title{findmarkers_gene2cell_mapping}
\usage{
findmarkers_gene2cell_mapping(
  obj,
  ref_genes_csv,
  findmarkers_csv,
  resolutions = seq(0.1, 1, by = 0.1),
  ref_gene_colname,
  ref_cell_colname,
  out_dir = glue("{getwd()}/FindMarkers_gene2cell_mapping"),
  assay_snn = "RNA_snn_res"
)
}
\description{
findmarkers_gene2cell_mapping
}
