% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_legend}
\alias{format_legend}
\title{format_legend}
\usage{
format_legend(
  my_plot,
  ...,
  point_size = 0.5,
  text_size = 5,
  space_legend = 0.2
)
}
\arguments{
\item{my_plot}{(ggplot obj): \cr
ggplot object to return}

\item{point_size}{(float): \cr
size to make points in legend}

\item{text_size}{(int or double): \cr
size to make text in legend}

\item{space_legend}{(float): \cr
linespacing of legend}
}
\value{
ggplot object
}
\description{
Ensure single-column, appropriately-sized legend in ggplot
}
\examples{
p1 <- format_legend(DimPlot(seurat_obj, reduction = "umap",
                   group.by = "sample", label = TRUE, label.size = 2,
                   repel = TRUE))
}
