% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{magick_overlay}
\alias{magick_overlay}
\title{magick_overlay}
\usage{
magick_overlay(
  main_plt,
  trans_plt,
  dest,
  filename,
  ...,
  x_dim = 14,
  y_dim = 17,
  transparency = 20,
  return_im = FALSE
)
}
\arguments{
\item{main_plt}{(ggplot obj): \cr ggplot to display with 0\% translucency}

\item{trans_plt}{(ggplot obj): \cr ggplot to display with \code{transparency}\%
translucency}

\item{dest}{(string or character vector): \cr directory or directories to
save overlayed magick image object (will create directories if not yet in
existence)}

\item{filename}{(string): \cr file name to save overlay plot to (include file
extension: "png" only)}

\item{x_dim}{(int or double = 14): \cr positive x axis limit for overlay
plot. \cr
Example: \cr
To create a plot with x values from [-17, 17], supply: \code{x_dim} = 17.
Assumed symmetric about the origin.}

\item{y_dim}{(int or double = 17): \cr positive y axis limit for overlay
plot. \cr
Example: \cr
To create a plot with y values from [-17, 17], supply \code{y_dim} = 17.
Assumed symmetric about the origin.}

\item{transparency}{(int or double = 20): \cr percent transparency to make
the transparent plot in the overlay}

\item{return_im}{(bool = FALSE) whether to return magick image object}
}
\value{
magick image object of overlayed plots (\code{main_plt} + \code{trans_plt}) if
\code{return_im} == TRUE. Else N/A
}
\description{
Create an overlay of two plots, with \code{transparency}\%
translucency, to visualize both simultaneously.
\cr
NOTE: Ideally, make ggtitle the same value on the two plots to overlay or
nonexistent. Additionally, best to supply plots without axes unless
identical prior to function call.
}
\examples{
im <- magick_overlay(main_plt = p1, trans_plt = p2,
                     dest = "~/endothelial/images/overlays/by_age",
                     filename = "10_PCW.png")
}
