% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_power_plots}
\alias{make_power_plots}
\title{make_power_plots}
\usage{
make_power_plots(
  cell_types,
  marker_dir,
  markerfile_command,
  outdir,
  ngenes = 30,
  plot_class = "power"
)
}
\arguments{
\item{cell_types}{(char vector): \cr
TODO}

\item{marker_dir}{(string): \cr
TODO}

\item{markerfile_command}{(string): \cr
TODO}

\item{outdir}{(string): \cr
TODO}

\item{ngenes}{(int = 30): \cr
TODO}

\item{plot_class}{(string = "power"): \cr
TODO}
}
\value{
N/A
}
\description{
Flush out
}
\examples{
 TODO
}
