% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_volc_plot}
\alias{make_volc_plot}
\title{make_volc_plot}
\usage{
make_volc_plot(
  data,
  title,
  lfc_color,
  lfc_label,
  xlab = "logFC",
  ylab = "p_adj.glb",
  sig_pval = 1.3
)
}
\arguments{
\item{data}{(seurat obj): \cr
TODO}

\item{title}{(string): \cr
TODO}

\item{lfc_color}{(float or int): \cr
TODO}

\item{lfc_label}{(float or int): \cr
TODO}

\item{xlab}{(string = "logFC"): \cr
TODO}
}
\value{
N/A
}
\description{
Flush out
}
\examples{
 TODO
}
