% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{query_integrated_reference}
\alias{query_integrated_reference}
\title{query_integrated_reference}
\usage{
query_integrated_reference(
  references,
  query,
  out_dir,
  npcs = 30,
  base_dir = getwd(),
  include_query = TRUE,
  return = TRUE,
  assay = NULL,
  reference_assay = "RNA",
  query_assay = "RNA",
  scale = FALSE,
  num_anchor_feats = 2000
)
}
\arguments{
\item{references}{(Seurat obj or list of Seurat objs): \cr dataset(s) to
integrate. \cr
NOTE: ensure \code{references} are normalized and variable features are found.
Only run \link[Seurat]{FindVariableFeatures} on unnormalized data.}

\item{query}{(Seurat obj): \cr  query dataset of cells to label}

\item{out_dir}{(str): \cr parent directory name to hold images and
integration objects created}

\item{npcs}{(int or double = 30): \cr number of principal components with
which to run \link[Seurat]{FindIntegrationAnchors},
\link[Seurat]{IntegrateData}, \link[Seurat]{RunUMAP},
\link[Seurat]{FindTransferAnchors}, and \link[Seurat]{TransferData}.}

\item{base_dir}{(string = \link{getwd()} (currrent working dir)
or char vect): \cr directory or directories) to save the plot in}

\item{include_query}{(bool = TRUE): \cr whether or not to include the \code{query}
dataset when building the reference atlas of cells and cell type labels to
compare \code{query} against. \cr
NOTE: Excluding `query`` will not allow your newly-assigned cell type
labels to include what they were originally, even if they would be found as
the ideal match by this workflow. All original cell type labels will be
replaced.}

\item{return}{(bool = TRUE): \cr if return = TRUE, return the query and
integrated Seurat objects created by this analysis.}

\item{assay}{(NULL or str): \cr find integration anchors for specified Seurat
object assay. If not provided, will use DefaultAssay of each reference (&
query, if include_query = TRUE) to integrate along.}
}
\value{
(named list): \cr list of named Seurat objects: \cr
query =  query dataset with newly-constructed cell type labels in assay
predicted.cell_type) \cr
integrated dataset
}
\description{
Integrate datasets (including or excluding \code{query} dataset) to
create a reference against which to assign cell type labels to the \code{query}
dataset. Datasets must be Seurat objects. Cell type labels must be under
slot "cell_type" for all datasets. \cr
Creates three directories (two subdirs of the first):  \cr
\itemize{
\item \code{out_dir}  \cr
\item \code{out_dir}/images  \cr
dir to hold plots created by this process. These plots are designed
to assess successful completion of the integration. \cr
\item \code{out_dir}/integration_objs
dir to hold all integration objects created in the process, in case
any step must be repeated or object(s) exported.
NOTES: Should have already run pca and other items on each object.
}
}
\examples{
output <- query_integrated_reference(references = list(seurat_obj1,
           seurat_obj2), query = seurat_obj3, npcs = 50, base_dir = "~/",
           include_query = FALSE, assay = "RNA", return = TRUE)
query <- output$query
integrated <- output$integrated
}
