% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemann.R
\name{riemann}
\alias{riemann}
\title{Implement Riemann Integration for a Generic Function}
\usage{
riemann(n, f, range)
}
\arguments{
\item{n}{The number of intervals used for Riemann approximation}

\item{f}{A function with a single argument x}

\item{range}{A vector containing the lower and upper endpoints of the range of integration}
}
\value{
Returns a single value which is the Riemann approximation of the function over (a,b)
}
\description{
Implement Riemann Integration for a Generic Function
}
\examples{
riemann(5,sqrt, c(1,3))
}
