% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save_fxn}
\alias{save_fxn}
\title{save_fxn}
\usage{
save_fxn(obj, file, dirs, ..., w = 8, h = 5)
}
\arguments{
\item{obj}{(ggplot obj): \cr
ggplot object to save}

\item{file}{(string): \cr
file name to save ggplot as (include file extension: "png" only)}

\item{dirs}{(string or character vector): \cr
directory (or directories) to save the plot in}

\item{w}{(int or double): \cr
width of PNG to save in inches}

\item{h}{(int or double): \cr
height of PNG to save in inches}
}
\value{
N/A
}
\description{
Save a ggplot object to the appropriate location(s)
}
\examples{
 save_fxn(main_plt, file = "tmp_plt1.png", dirs = c("~/endothelial/images",
        "~/examples/images"))
}
