% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid.R
\name{trapezoid}
\alias{trapezoid}
\title{Implement Trapezoid Rule for Generic Function}
\usage{
trapezoid(n, f, range)
}
\arguments{
\item{n}{The number of intervals used for approximation}

\item{f}{A function with a single argument x}

\item{range}{A vector containing the lower and upper endpoints of the range of integration}
}
\value{
Returns a single value which is the trapezoidal approximation of the function over (a,b)
}
\description{
Implement Trapezoid Rule for Generic Function
}
\examples{
trapezoid(sqrt, c(1,3),5)

}
